% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_status.R
\name{sum_wrapper}
\alias{sum_wrapper}
\alias{sum_wrapper.logical}
\alias{sum_wrapper.factor}
\alias{sum_wrapper.numeric}
\alias{sum_wrapper.Date}
\alias{sum_wrapper.character}
\title{Sum wrapper}
\usage{
sum_wrapper(x, ...)

\method{sum_wrapper}{logical}(y, by)

\method{sum_wrapper}{factor}(y, by)

\method{sum_wrapper}{numeric}(y, by)

\method{sum_wrapper}{Date}(y, by)

\method{sum_wrapper}{character}(y, by)
}
\arguments{
\item{x}{any vector}

\item{...}{other arguments}
}
\description{
summary wrapper function
}
\details{
A details of function
}
