% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_status.R
\name{summary_table}
\alias{summary_table}
\title{Sum table}
\usage{
summary_table(z, by = NULL)
}
\arguments{
\item{z}{a data.frame}

\item{by}{variable name for rates}
}
\description{
Summary table wrapper function
}
\details{
A details of function
}
