% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yll_cox.R
\name{yll_ci}
\alias{yll_ci}
\title{YLL CI}
\usage{
yll_ci(
  model,
  data,
  newdata,
  predvar,
  tau,
  conf_int = 0.95,
  iterations = 1000,
  multicore = FALSE,
  ncpus = 1
)
}
\arguments{
\item{model}{coxph model}

\item{data}{data used to fit model}

\item{newdata}{newdata same as YLL}

\item{tau}{max observation length}

\item{conf_int}{Confidence interval}

\item{iterations}{default 1000}

\item{multicore}{boolean default FALSE}

\item{ncpus}{number of cpus to use}
}
\description{
Calculate bootstraped CI
}
