% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yll_cox.R
\name{yll_cox}
\alias{yll_cox}
\title{Years of life lost calculation from coxph model}
\usage{
yll_cox(
  model,
  data,
  newdata,
  predvar = NULL,
  tau = 18,
  ci.level = NULL,
  iterations = 1000,
  multicore = FALSE,
  ncpus = 1
)
}
\arguments{
\item{model}{coxph model}

\item{data}{data used to fit coxph model}

\item{newdata}{New data at which to do predictions}

\item{predvar}{variables to predict yll for}

\item{tau}{Maximum life length}

\item{ci.level}{Confidence level, default set to 0.95}

\item{iterations}{default 1000}

\item{multicore}{boolean default FALSE}

\item{ncpus}{number of cpus to use}
}
\description{
Years of life lost calculation from coxph model
}
\examples{
library(survival)

mdl <- coxph(Surv(age_end, event)~ gender*dis + ses, data = popum_dis)

newdata <- with(popum_dis, expand.grid(
  gender = unique(gender)[1],
  ses = unique(ses)[1],
  dis = unique(dis)
))


yll_cox(mdl, popum_dis, newdata = newdata, predvar = c("dis", "gender"), tau = 43)

\dontrun{
yll_cox(
  mdl,
  popum_dis,
  newdata = newdata,
  tau = 43,
  ci.level = .95,
  iterations = 500,
  multicore = T,
  ncpus = 11
)
}
}
